修改传webhook细节
多传一点数据给我嘛

+++ Core/RuntimeObject/Danmu.cs
@@ -438,7 +438,7 @@ namespace Core.RuntimeObject
             /// <summary>
             /// 花费：单位金瓜子
             /// </summary>
-            public float Price { set; get; }
+            public double Price { set; get; }
             /// <summary>
             /// 礼物名称
             /// </summary>
